refactor: Auth 페이지 경로 추가
