+++ src/pages/Auth/SignUp/SignUp.style.ts
