+++ src/pages/Auth/SignUp/components/EmailVerifyModal/EmailVerifyModal.style.ts
