+++ src/pages/Auth/SignUpCheck/SignUpCheck.style.ts
