+++ src/pages/Auth/SignUpCheck/SignUpCheck.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import * as S from './SignupCheck.style.ts';
+import * as S from './SignUpCheck.style.ts';
 import LogoSvg from '@/assets/AuthImg/AuthLogo.svg';
 import { PRIVACY_POLICY } from '@/constants/policy';
