+++ src/pages/Auth/SignUpGoogle/SignUpGoogle.style.ts
